refactor(core): extract maintenance/backfill-tags.ts — tags_text backfill
